requirements.txt
Flask
gunicorn
sqlalchemy
pymysql
openai
pygpt4all
gTTS
python-dotenv
pygame